refactor: Sorting out some files into animation directory.
