+++ CMakeLists.txt
@@ -13,42 +13,47 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/cmake/")
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 set(PUBLIC_HEADERS
-	include/moth_ui/animation_clip.h
-	include/moth_ui/animation_controller.h
-	include/moth_ui/animation_track.h
+	include/moth_ui/animation/animation_clip_controller.h                                                                                                                                         
+	include/moth_ui/animation/animation_clip.h                                                                                                                                                    
+	include/moth_ui/animation/animation_controller.h                                                                                                                                              
+	include/moth_ui/animation/animation_event.h                                                                                                                                                   
+	include/moth_ui/animation/animation_track_controller.h                                                                                                                                        
+	include/moth_ui/animation/animation_track.h                                                                                                                                                   
+	include/moth_ui/animation/clip_controller.h                                                                                                                                                   
+	include/moth_ui/animation/keyframe.h                                                                                                                                                          
 	include/moth_ui/blend_mode.h
-	include/moth_ui/clip_controller.h
 	include/moth_ui/context.h
-	include/moth_ui/events/event.h
-	include/moth_ui/events/event_animation.h
-	include/moth_ui/events/event_key.h
-	include/moth_ui/events/event_mouse.h
 	include/moth_ui/event_dispatch.h
 	include/moth_ui/event_listener.h
+	include/moth_ui/events/event_animation.h
+	include/moth_ui/events/event.h
+	include/moth_ui/events/event_key.h
+	include/moth_ui/events/event_mouse.h
 	include/moth_ui/font_factory.h
 	include/moth_ui/group.h
+	include/moth_ui/ifont_factory.h
 	include/moth_ui/ifont.h
-	include/moth_ui/iimage.h
 	include/moth_ui/iimage_factory.h
+	include/moth_ui/iimage.h
 	include/moth_ui/image_scale_type.h
 	include/moth_ui/irenderer.h
 	include/moth_ui/itarget.h
 	include/moth_ui/layers/layer.h
 	include/moth_ui/layers/layer_stack.h
-	include/moth_ui/layout/layout.h
 	include/moth_ui/layout/layout_cache.h
-	include/moth_ui/layout/layout_entity.h
 	include/moth_ui/layout/layout_entity_clip.h
 	include/moth_ui/layout/layout_entity_group.h
+	include/moth_ui/layout/layout_entity.h
 	include/moth_ui/layout/layout_entity_image.h
 	include/moth_ui/layout/layout_entity_rect.h
 	include/moth_ui/layout/layout_entity_ref.h
 	include/moth_ui/layout/layout_entity_text.h
 	include/moth_ui/layout/layout_entity_type.h
+	include/moth_ui/layout/layout.h
 	include/moth_ui/layout/layout_rect.h
-	include/moth_ui/node.h
 	include/moth_ui/node_clip.h
 	include/moth_ui/node_factory.h
+	include/moth_ui/node.h
 	include/moth_ui/node_image.h
 	include/moth_ui/node_rect.h
 	include/moth_ui/node_text.h
@@ -64,26 +69,29 @@ set(PUBLIC_HEADERS
 	include/moth_ui/utils/vector_serialization.h
 	include/moth_ui/widget.h
 )
-
 set(SOURCES ${SOURCES}
-	src/animation_controller.cpp
-	src/animation_track.cpp
+	src/animation/animation_clip_controller.cpp
+	src/animation/animation_controller.cpp
+	src/animation/animation_event.cpp
+	src/animation/animation_track_controller.cpp
+	src/animation/animation_track.cpp
+	src/animation/keyframe.cpp
 	src/context.cpp
 	src/font_factory.cpp
 	src/group.cpp
 	src/layers/layer.cpp
 	src/layers/layer_stack.cpp
-	src/layout/layout.cpp
 	src/layout/layout_cache.cpp
-	src/layout/layout_entity.cpp
+	src/layout/layout.cpp
 	src/layout/layout_entity_clip.cpp
+	src/layout/layout_entity.cpp
 	src/layout/layout_entity_group.cpp
 	src/layout/layout_entity_image.cpp
 	src/layout/layout_entity_rect.cpp
 	src/layout/layout_entity_ref.cpp
 	src/layout/layout_entity_text.cpp
-	src/node.cpp
 	src/node_clip.cpp
+	src/node.cpp
 	src/node_factory.cpp
 	src/node_image.cpp
 	src/node_rect.cpp
