+++ include/moth_ui/animation/animation_clip.h
