+++ include/moth_ui/animation/animation_clip_controller.h
