+++ include/moth_ui/animation/animation_track_controller.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "moth_ui/animation_track.h"
+#include "moth_ui/animation/animation_track.h"
 #include "moth_ui/ui_fwd.h"
 
 namespace moth_ui {
