+++ include/moth_ui/layout/layout_entity.h
@@ -2,7 +2,7 @@
 
 #include "moth_ui/context.h"
 #include "moth_ui/ui_fwd.h"
-#include "moth_ui/animation_track.h"
+#include "moth_ui/animation/animation_track.h"
 #include "moth_ui/layout/layout_entity_type.h"
 #include "moth_ui/utils/color.h"
 #include "moth_ui/blend_mode.h"
