+++ src/animation/animation_clip_controller.cpp
@@ -1,7 +1,7 @@
 #include "common.h"
-#include "moth_ui/animation_clip_controller.h"
-#include "moth_ui/animation_controller.h"
-#include "moth_ui/animation_clip.h"
+#include "moth_ui/animation/animation_clip_controller.h"
+#include "moth_ui/animation/animation_controller.h"
+#include "moth_ui/animation/animation_clip.h"
 #include "moth_ui/events/event_animation.h"
 #include "moth_ui/group.h"
 #include "moth_ui/layout/layout_entity_group.h"
