+++ src/layout/layout_entity_ref.cpp
@@ -2,7 +2,7 @@
 #include "moth_ui/layout/layout_entity_ref.h"
 #include "moth_ui/group.h"
 #include "moth_ui/layout/layout.h"
-#include "moth_ui/animation_clip.h"
+#include "moth_ui/animation/animation_clip.h"
 
 namespace moth_ui {
     LayoutEntityRef::LayoutEntityRef(LayoutRect const& initialBounds, Layout const& layoutRef)
