+++ src/node.cpp
@@ -1,7 +1,7 @@
 #include "common.h"
 #include "moth_ui/node.h"
 #include "moth_ui/layout/layout_entity.h"
-#include "moth_ui/animation_controller.h"
+#include "moth_ui/animation/animation_controller.h"
 #include "moth_ui/group.h"
 #include "moth_ui/context.h"
